+++ catalog/app/containers/Bucket/Dir/Toolbar/Get/Options.tsx
@@ -2,9 +2,11 @@ import { basename } from 'path'
 
 import * as React from 'react'
 import * as M from '@material-ui/core'
+import * as Icons from '@material-ui/icons'
 
 import * as CodeSamples from 'containers/Bucket/CodeSamples'
 import * as Buttons from 'containers/Bucket/Download/Buttons'
+import { ZipDownloadForm } from 'containers/Bucket/FileView'
 import GetOptions from 'containers/Bucket/Toolbar/GetOptions'
 import type * as Toolbar from 'containers/Bucket/Toolbar'
 import type { Features } from '../useFeatures'
@@ -35,20 +37,45 @@ function DirCodeSamples({ className, bucket, path }: DirCodeSamplesProps) {
   )
 }
 
+const useDownloadDirStyles = M.makeStyles((t) => ({
+  root: {
+    marginBottom: t.spacing(1),
+  },
+  group: {
+    whiteSpace: 'nowrap',
+    width: '100%',
+  },
+  main: {
+    flexGrow: 1,
+    flexShrink: 0,
+    justifyContent: 'flex-start',
+  },
+}))
+
 interface DownloadDirProps {
   dirHandle: Toolbar.DirHandle
 }
 
 function DownloadDir({ dirHandle }: DownloadDirProps) {
-  // TODO: pass selection to Buttons.DownloadDir
+  const classes = useDownloadDirStyles()
   const feedback = Buttons.useDownloadFeedback()
   return (
-    <Buttons.DownloadDir
+    <ZipDownloadForm
+      className={classes.root}
       suffix={`dir/${dirHandle.bucket}/${dirHandle.path}`}
-      {...feedback}
     >
-      Download ZIP (directory)
-    </Buttons.DownloadDir>
+      <M.ButtonGroup variant="outlined" className={classes.group}>
+        <M.Button
+          className={classes.main}
+          startIcon={<Icons.ArchiveOutlined />}
+          type="submit"
+          {...feedback}
+        >
+          Download ZIP (directory)
+        </M.Button>
+        <Buttons.CopyButton uri={`s3://${dirHandle.bucket}/${dirHandle.path}`} />
+      </M.ButtonGroup>
+    </ZipDownloadForm>
   )
 }